backend/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
contextlib2==21.6.0
cryptography==45.0.6
diff-match-patch==20241021
Django==5.2.5
django-cors-headers==4.7.0
django-import-export==4.3.9
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
email_validator==2.2.0
fire==0.7.1
idna==3.10
injector==0.22.0
markdown-it-py==4.0.0
mdurl==0.1.2
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyYAML==6.0.2
qrcode==8.2
rav==0.0.9
requests==2.32.4
rich==14.1.0
segno==1.6.6
sqlparse==0.5.3
tablib==3.8.0
termcolor==3.1.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
